Refactor infrastructure persistense

+++ Aero.Domain/Entities/Sensor.cs
@@ -4,8 +4,8 @@ namespace Aero.Domain.Entities;
 
 public sealed class Sensor : BaseDomain
 {
-      public short ModuleId { get; set; }
-    public short DoorId { get; set; }
+      public int ModuleId { get; set; }
+    public int DoorId { get; set; }
         public short InputNo { get; set; }
         public short InputMode { get; set; }
         public short Debounce { get; set; }
