+++ Aero.Infrastructure/Persistences/Entities/Sensor.cs
@@ -6,16 +6,16 @@ namespace Aero.Infrastructure.Persistences.Entities
     public sealed class Sensor : BaseEntity
     {
         public Module module { get; set; }
-        public short door_id { get; set; }
+        public int door_id { get; set; }
         public Door sensor_door { get; set; }
-        public short module_id { get; set; }
+        public int module_id { get; set; }
         public short input_no { get; set; }
         public short input_mode { get; set; }
         public short debounce { get; set; }
         public short holdtime { get; set; }
         public short dc_held { get; set; }
 
-        public Sensor(short module,short doorid,short inputno,short inputmode,short debounce,short holdtime,short dcheld,int location) : base(location) 
+        public Sensor(int module,int doorid,short inputno,short inputmode,short debounce,short holdtime,short dcheld,int location) : base(location) 
         {
             this.module_id = module;
             this.door_id = doorid;
