+++ Aero.Infrastructure/Persistences/Entities/Strike.cs
@@ -8,7 +8,7 @@ namespace Aero.Infrastructure.Persistences.Entities
         public int module_id { get; set; }
         public Module module { get; set; }
         public Door strike_door { get; set; }
-        public short door_id { get; set; }
+        public int door_id { get; set; }
         public short output_no { get; set; }
         public short relay_mode { get; set; }
         public short offline_mode { get; set; }
@@ -16,7 +16,7 @@ namespace Aero.Infrastructure.Persistences.Entities
         public short strike_min { get; set; }
         public short strike_mode { get; set; }
 
-        public Strike(int moduleid,short doorid,short outputno,short relaymode,short offlinemode,short strikmax,short strikemin,short strikemode,int location) : base(location) 
+        public Strike(int moduleid,int doorid,short outputno,short relaymode,short offlinemode,short strikmax,short strikemin,short strikemode,int location) : base(location) 
         {
             this.module_id = moduleid;
             this.door_id = doorid;
